use LargeDatasetSampler

+++ primus_dlrm/data/counters.py
@@ -146,7 +146,7 @@ def precompute_user_counters(
     counter_windows_days = list(counter_windows_days)
     N = flat_uid.shape[0]
     W = len(counter_windows_days)
-    out = np.zeros((N, 3 * W), dtype=np.float16)
+    out = np.zeros((N, 3 * W), dtype=np.float32)
 
     seg_s, seg_e = _compute_segments(flat_uid)
     timestamps = flat_ts.astype(np.int64, copy=False)
@@ -171,7 +171,7 @@ def precompute_user_counters(
             f"  user_counters[{w_days}d]: {time.time() - t0:.1f}s "
             f"({len(seg_s):,} segments, {N:,} events)"
         )
-    out[:] = out64.astype(np.float16, copy=False)
+    out[:] = out64.astype(np.float32, copy=False)
     return out
